SSPCPP-691 replace ignoreCase with caseSensitive
https://issues.shibboleth.net/jira/browse/SSPCPP-691

XMLAccessControl is relatively uncontroversion.

XMLRequestMapper is harder - we need to maintain the legacy broken behavior whereby
'ignoreCase' means 'caseSensitive' in the PathRegex case.

So we for a PathRegex we issue an ERROR on ignoreCase and give it caseSensitive
behavior (backwards compatibility).

We do the right thing for caseSensitivity.

We default to caseSensitivity = false (a change in default, but one which fixes
legacy misbehavior).