SSPCPP-682 Wire Shibd into event messages
https://issues.shibboleth.net/jira/browse/SSPCPP-682

Use specific messages for shibd.  No formatting is needed, but as
a matter of course we leave the original message in place (to help
debug in the case of broken installs) and also put GetLastError()
into the binary data area.

We cannot format binary data in the event viewer, so one would
need to look at tht raw xml and then byte swap. But that is
advanced debugging anyway.