SSPCPP-682: Messages and plumbing for ISAPI Filter
https://issues.shibboleth.net/jira/browse/SSPCPP-682

Convert the use of the event log to exploit a message file.  In
many cases we use parameterization.  In other we just pass the
old string in as an aid to debugging borken installs.  In all cases
we plug GetLastError into the binary data.

WriteClientError (two variants) which writes to the log gets an
extra parameters which allows for better logging, and indeed to
supress the error log if we have written more data to it already.