SSPCPP-604 IIS7 Native Support for IIS7
https://issues.shibboleth.net/jira/browse/SSPCPP-694

This initial commit captures all the work described in the case notes to date.
This has to be hand installed and all the DLLs colocated in the IIS directory.

It is reasonably fully functional but there are several todos in the code
and significant further test and nature study is required.