SSPCPP-694 Prototype roles-based Authn
https://issues.shibboleth.net/jira/browse/SSPCPP-694

Prototype up how to add roles in an IIS plugin. Based on this prototypical config

<IIS normalizeRequest="true" safeHeaderNames="true" useHeaders="false" useVariables="true">
  <Roles authNRole="ShibAuthenticated">
    <Role attribute="epa" prefix="AffiliatedAs"/>
  </Roles>
</IIS>

Populate the principal so that it has the role "ShibAuthenticated" and the roles consisting of every value provided to epa appropriated prefixed.  We can then do standard roles based AuthZ in the IIS GUI.

The precisely plumbin is questionable (tokenising the setHeader() parameterization), but it keeps things non-intrusive for now.