Strip NO_START functionality from the init script
Quoting Lintian:

This is an older practice used so that the package's init script would
not start the service until the local system administrator changed this
value.

However, this hides from the underlying init system whether or not the
daemon should actually be started leading to confusing behavior
including "service package start" returning success without the service
actually starting.